Fix 32-bit integer overflow in timing code and switch sfincs_lib to omp_get_wtime()

+++ source/src/sfincs_continuity.f90
@@ -11,10 +11,10 @@ contains
    real*4           :: dt
    real*8           :: t
    !
-   integer          :: count0
-   integer          :: count1
-   integer          :: count_rate
-   integer          :: count_max
+   integer*8        :: count0
+   integer*8        :: count1
+   integer*8        :: count_rate
+   integer*8        :: count_max
    real             :: tloop
    !
    call system_clock(count0, count_rate, count_max)
