+++ source/src/sfincs_discharges.f90
@@ -318,10 +318,10 @@ contains
    !
    implicit none
    !
-   integer  :: count0
-   integer  :: count1
-   integer  :: count_rate
-   integer  :: count_max
+   integer*8 :: count0
+   integer*8 :: count1
+   integer*8 :: count_rate
+   integer*8 :: count_max
    real     :: tloop
    !
    real*8           :: t
